core: Add internal helper API for adding query strings
In the Drive v2 API, certain GDataDocumentsQuery properties like
show-deleted and show-folders no longer have their own parameters, but
have to be specified as a search clause in the query string. This will
be used internally by child classes of GDataQuery to add search clauses
that represent service-specific query properties.

https://bugzilla.gnome.org/show_bug.cgi?id=684920